Split up into more nuget packages.
Mostly moving stuff out of web
