+++ src/Pact.EntityFrameworkCore/README.md
@@ -1,0 +1,2 @@
+# Pact.EntityFrameworkCore 🕸
+Provides a number of useful extension methods for Entity Framework Core.
