+++ test/Pact.Core.Tests/Containers/Basic.cs
@@ -1,0 +1,9 @@
+﻿namespace Pact.Core.Tests.Containers
+{
+    public class Basic
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public bool SoftDelete { get; set; }
+    }
+}
